+++ modules/core/src/main/scala/io/bartholomews/discogs4s/DiscogsSimpleClient.scala
@@ -52,6 +52,16 @@ class DiscogsSimpleClient[F[_], S <: Signer] private[discogs4s] (client: FsClien
         responseHandler: ResponseHandler[DE, PaginatedReleases]
     ): F[SttpResponse[DE, PaginatedReleases]] =
       api.getArtistReleases(artistId, sortBy, sortOrder)(client.signer)
+
+    def getLabel[DE](labelId: Label.Id)(implicit
+        responseHandler: ResponseHandler[DE, Label]
+    ): F[SttpResponse[DE, Label]] =
+      api.getLabel(labelId)(client.signer)
+
+    def getLabelReleases[DE](labelId: Label.Id, page: Int = 1, perPage: Int = 50)(implicit
+        responseHandler: ResponseHandler[DE, Release]
+    ): F[SttpResponse[DE, Release]] =
+      api.getLabelReleases(labelId, page, perPage)(client.signer)
   }
 
   final object users {