+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `DatabaseApi.getLabel`
+- `DatabaseApi.getLabelReleases`
 
 ## [0.1.2] - 2021-08-05
 ### Added
@@ -14,11 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Multi-module projects setup (circe / play / core)
 - Multi-client setup (for each auth flow)
 - Improved types on `UserIdentity` and `UserProfile` (breaking changes)
-  
+
 ## [0.1.1] - 2021-01-02
 ### Changed
 - Fsclient `client` val in `DiscogsClient` is now accessible,
-  so its signer can be used for *Personal access token* calls 
+  so its signer can be used for *Personal access token* calls
 
 ## [0.1.0] - 2021-01-02
 ### Added